pyproject.toml
[project]
name = "farmbros-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "geoalchemy2>=0.18.0",
    "psycopg2>=2.9.10",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.8",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
    "werkzeug>=3.1.3",
]
